+++ pyroute2/netlink/rtnl/ifinfmsg.py
@@ -1,8 +1,67 @@
+import os
+import time
+import json
+import struct
+import platform
+import subprocess
+from fcntl import ioctl
 from pyroute2.common import map_namespace
+from pyroute2.common import ANCIENT
+# from pyroute2.netlink import NLMSG_ERROR
 from pyroute2.netlink import nla
 from pyroute2.netlink import nlmsg
 from pyroute2.netlink.rtnl.iw_event import iw_event
 
+
+# it's simpler to double constants here, than to change all the
+# module layout; but it is a subject of the future refactoring
+RTM_NEWLINK = 16
+RTM_DELLINK = 17
+#
+
+_ANCIENT_BARRIER = 0.3
+_BONDING_MASTERS = '/sys/class/net/bonding_masters'
+_BONDING_SLAVES = '/sys/class/net/%s/bonding/slaves'
+_BRIDGE_MASTER = '/sys/class/net/%s/brport/bridge/ifindex'
+_BONDING_MASTER = '/sys/class/net/%s/master/ifindex'
+IFNAMSIZ = 16
+
+TUNDEV = '/dev/net/tun'
+arch = platform.machine()
+if arch == 'x86_64':
+    TUNSETIFF = 0x400454ca
+    TUNSETPERSIST = 0x400454cb
+    TUNSETOWNER = 0x400454cc
+    TUNSETGROUP = 0x400454ce
+elif arch == 'ppc64':
+    TUNSETIFF = 0x800454ca
+    TUNSETPERSIST = 0x800454cb
+    TUNSETOWNER = 0x800454cc
+    TUNSETGROUP = 0x800454ce
+else:
+    TUNSETIFF = None
+
+##
+#
+# tuntap flags
+#
+IFT_TUN = 0x0001
+IFT_TAP = 0x0002
+IFT_NO_PI = 0x1000
+IFT_ONE_QUEUE = 0x2000
+IFT_VNET_HDR = 0x4000
+IFT_TUN_EXCL = 0x8000
+IFT_MULTI_QUEUE = 0x0100
+IFT_ATTACH_QUEUE = 0x0200
+IFT_DETACH_QUEUE = 0x0400
+# read-only
+IFT_PERSIST = 0x0800
+IFT_NOFILTER = 0x1000
+
+##
+#
+# normal flags
+#
 IFF_UP = 0x1  # interface is up
 IFF_BROADCAST = 0x2  # broadcast address valid
 IFF_DEBUG = 0x4  # turn on debugging
@@ -76,7 +135,7 @@ stats_names = ('rx_packets',
                'tx_compressed')
 
 
-class ifinfmsg(nlmsg):
+class ifinfbase(object):
     '''
     Network interface message
     struct ifinfomsg {
@@ -115,7 +174,7 @@ class ifinfmsg(nlmsg):
                ('IFLA_OPERSTATE', 'state'),
                ('IFLA_LINKMODE', 'uint8'),
                ('IFLA_LINKINFO', 'ifinfo'),
-               ('IFLA_NET_NS_PID', 'hex'),
+               ('IFLA_NET_NS_PID', 'uint32'),
                ('IFLA_IFALIAS', 'hex'),
                ('IFLA_NUM_VF', 'uint32'),
                ('IFLA_VFINFO_LIST', 'hex'),
@@ -124,7 +183,7 @@ class ifinfmsg(nlmsg):
                ('IFLA_PORT_SELF', 'hex'),
                ('IFLA_AF_SPEC', 'af_spec'),
                ('IFLA_GROUP', 'uint32'),
-               ('IFLA_NET_NS_FD', 'hex'),
+               ('IFLA_NET_NS_FD', 'netns_fd'),
                ('IFLA_EXT_MASK', 'hex'),
                ('IFLA_PROMISCUITY', 'uint32'),
                ('IFLA_NUM_TX_QUEUES', 'uint32'),
@@ -157,7 +216,39 @@ class ifinfmsg(nlmsg):
         # convert flags
         if isinstance(self['flags'], (set, tuple, list)):
             self['flags'], self['change'] = self.names2flags(self['flags'])
-        return nlmsg.encode(self)
+        return super(ifinfbase, self).encode()
+
+    class netns_fd(nla):
+        fields = [('value', 'I')]
+        netns_run_dir = '/var/run/netns'
+        netns_fd = None
+
+        def encode(self):
+            self.close()
+            #
+            # There are two ways to specify netns
+            #
+            # 1. provide fd to an open file
+            # 2. provide a file name
+            #
+            # In the first case, the value is passed to the kernel
+            # as is. In the second case, the object opens appropriate
+            # file from `self.netns_run_dir` and closes it upon
+            # `__del__(self)`
+            if isinstance(self.value, int):
+                self['value'] = self.value
+            else:
+                self.netns_fd = os.open('%s/%s' % (self.netns_run_dir,
+                                                   self.value), os.O_RDONLY)
+                self['value'] = self.netns_fd
+            nla.encode(self)
+
+        def __del__(self):
+            self.close()
+
+        def close(self):
+            if self.netns_fd is not None:
+                os.close(self.netns_fd)
 
     class wireless(iw_event):
         pass
@@ -207,7 +298,41 @@ class ifinfmsg(nlmsg):
                 return self.vlan_data
             elif kind == 'bond':
                 return self.bond_data
+            elif kind == 'veth':
+                return self.veth_data
+            elif kind == 'tuntap':
+                return self.tuntap_data
+            elif kind == 'bridge':
+                return self.bridge_data
             return self.hex
+
+        class tuntap_data(nla):
+            '''
+            Fake data type
+            '''
+            prefix = 'IFTUN_'
+
+            nla_map = (('IFTUN_UNSPEC', 'none'),
+                       ('IFTUN_MODE', 'asciiz'),
+                       ('IFTUN_UID', 'uint32'),
+                       ('IFTUN_GID', 'uint32'),
+                       ('IFTUN_IFR', 'flags'))
+
+            class flags(nla):
+                fields = (('no_pi', 'B'),
+                          ('one_queue', 'B'),
+                          ('vnet_hdr', 'B'),
+                          ('tun_excl', 'B'),
+                          ('multi_queue', 'B'),
+                          ('persist', 'B'),
+                          ('nofilter', 'B'))
+
+        class veth_data(nla):
+            nla_map = (('VETH_INFO_UNSPEC', 'none'),
+                       ('VETH_INFO_PEER', 'info_peer'))
+
+            def info_peer(self, *argv, **kwarg):
+                return ifinfveth
 
         class vlan_data(nla):
             nla_map = (('IFLA_VLAN_UNSPEC', 'none'),
@@ -220,7 +345,13 @@ class ifinfmsg(nlmsg):
                 fields = (('flags', 'I'),
                           ('mask', 'I'))
 
+        class bridge_data(nla):
+            prefix = 'IFLA_BRIDGE_'
+            nla_map = (('IFLA_BRIDGE_STP_STATE', 'uint32'),
+                       ('IFLA_BRIDGE_MAX_AGE', 'uint32'))
+
         class bond_data(nla):
+            prefix = 'IFLA_BOND_'
             nla_map = (('IFLA_BOND_UNSPEC', 'none'),
                        ('IFLA_BOND_MODE', 'uint8'),
                        ('IFLA_BOND_ACTIVE_SLAVE', 'uint32'),
@@ -309,7 +440,8 @@ class ifinfmsg(nlmsg):
                        ('IFLA_INET6_MCAST', 'hex'),
                        ('IFLA_INET6_CACHEINFO', 'ipv6_cache_info'),
                        ('IFLA_INET6_ICMP6STATS', 'icmp6_stats'),
-                       ('IFLA_INET6_TOKEN', 'ip6addr'))
+                       ('IFLA_INET6_TOKEN', 'ip6addr'),
+                       ('IFLA_INET6_ADDR_GEN_MODE', 'uint8'))
 
             class ipv6_devconf(nla):
                 # ./include/uapi/linux/ipv6.h
@@ -393,3 +525,425 @@ class ifinfmsg(nlmsg):
                           ('outmsgs', 'Q'),
                           ('outerrors', 'Q'),
                           ('inmsgs', 'Q'))
+
+
+class ifinfmsg(ifinfbase, nlmsg):
+    pass
+
+
+class ifinfveth(ifinfbase, nla):
+    pass
+
+
+def proxy_linkinfo(data, nl):
+    offset = 0
+    inbox = []
+    while offset < len(data):
+        msg = ifinfmsg(data[offset:])
+        msg.decode()
+        inbox.append(msg)
+        offset += msg['header']['length']
+
+    data = b''
+    for msg in inbox:
+        kind = None
+        ifname = msg.get_attr('IFLA_IFNAME')
+
+        # fix master
+        if ANCIENT:
+            master = compat_get_master(ifname)
+            if master is not None:
+                msg['attrs'].append(['IFLA_MASTER', master])
+
+        # fix linkinfo & kind
+        li = msg.get_attr('IFLA_LINKINFO')
+        if li is not None:
+            kind = li.get_attr('IFLA_INFO_KIND')
+            if kind is None:
+                kind = get_interface_type(ifname)
+                li['attrs'].append(['IFLA_INFO_KIND', kind])
+        else:
+            kind = get_interface_type(ifname)
+            msg['attrs'].append(['IFLA_LINKINFO',
+                                 {'attrs': [['IFLA_INFO_KIND', kind]]}])
+
+        li = msg.get_attr('IFLA_LINKINFO')
+        # fetch specific interface data
+        if (kind in ('bridge', 'bond')) and \
+                [x for x in li['attrs'] if x[0] == 'IFLA_INFO_DATA']:
+            if kind == 'bridge':
+                t = '/sys/class/net/%s/bridge/%s'
+                ifdata = ifinfmsg.ifinfo.bridge_data
+            elif kind == 'bond':
+                t = '/sys/class/net/%s/bonding/%s'
+                ifdata = ifinfmsg.ifinfo.bond_data
+
+            commands = []
+            for cmd, _ in ifdata.nla_map:
+                try:
+                    with open(t % (ifname, ifdata.nla2name(cmd)), 'r') as f:
+                        value = f.read()
+                    if cmd == 'IFLA_BOND_MODE':
+                        value = value.split()[1]
+                    commands.append([cmd, int(value)])
+                except:
+                    pass
+            if commands:
+                li['attrs'].append(['IFLA_INFO_DATA', {'attrs': commands}])
+
+        msg.reset()
+        msg.encode()
+        data += msg.buf.getvalue()
+
+    return {'verdict': 'forward',
+            'data': data}
+
+
+def proxy_setlink(data, nl):
+
+    def get_interface(index):
+        msg = nl.get_links(index)[0]
+        return {'ifname': msg.get_attr('IFLA_IFNAME'),
+                'master': msg.get_attr('IFLA_MASTER'),
+                'kind': msg.
+                get_attr('IFLA_LINKINFO').
+                get_attr('IFLA_INFO_KIND')}
+
+    msg = ifinfmsg(data)
+    msg.decode()
+    forward = True
+
+    kind = None
+    infodata = None
+
+    ifname = msg.get_attr('IFLA_IFNAME') or \
+        get_interface(msg['index'])['ifname']
+    linkinfo = msg.get_attr('IFLA_LINKINFO')
+    if linkinfo:
+        kind = linkinfo.get_attr('IFLA_INFO_KIND')
+        infodata = linkinfo.get_attr('IFLA_INFO_DATA')
+
+    if kind in ('bond', 'bridge'):
+        code = 0
+        #
+        if kind == 'bond':
+            func = compat_set_bond
+        elif kind == 'bridge':
+            func = compat_set_bridge
+        #
+        for (cmd, value) in infodata.get('attrs', []):
+            cmd = infodata.nla2name(cmd)
+            code = func(ifname, cmd, value) or code
+        #
+        if code:
+            err = OSError()
+            err.errno = code
+            raise err
+
+    # is it a port setup?
+    master = msg.get_attr('IFLA_MASTER')
+    if master is not None:
+
+        if master == 0:
+            # port delete
+            # 1. get the current master
+            iface = get_interface(msg['index'])
+            master = get_interface(iface['master'])
+            cmd = 'del'
+        else:
+            # port add
+            # 1. get the master
+            master = get_interface(master)
+            cmd = 'add'
+
+        # 2. delete the port
+        if master['kind'] == 'team':
+            forward = manage_team_port(cmd, master['ifname'], ifname)
+        elif master['kind'] == 'bridge':
+            forward = compat_bridge_port(cmd, master['ifname'], ifname)
+        elif master['kind'] == 'bond':
+            forward = compat_bond_port(cmd, master['ifname'], ifname)
+
+    if forward is not None:
+        return {'verdict': 'forward',
+                'data': data}
+
+
+def proxy_dellink(data, nl):
+    orig_msg = ifinfmsg(data)
+    orig_msg.decode()
+
+    # get full interface description
+    msg = nl.get_links(orig_msg['index'])[0]
+    msg['header']['type'] = orig_msg['header']['type']
+
+    # get the interface kind
+    kind = None
+    li = msg.get_attr('IFLA_LINKINFO')
+    if li is not None:
+        kind = li.get_attr('IFLA_INFO_KIND')
+
+    if kind in ('ovs-bridge', 'openvswitch'):
+        return manage_ovs(msg)
+
+    if ANCIENT and kind in ('bridge', 'bond'):
+        # route the request
+        if kind == 'bridge':
+            compat_del_bridge(msg.get_attr('IFLA_IFNAME'))
+        elif kind == 'bond':
+            compat_del_bond(msg.get_attr('IFLA_IFNAME'))
+        # while RTM_NEWLINK is not intercepted -- sleep
+        time.sleep(_ANCIENT_BARRIER)
+        return
+
+    return {'verdict': 'forward',
+            'data': data}
+
+
+def proxy_newlink(data, nl):
+    msg = ifinfmsg(data)
+    msg.decode()
+    kind = None
+
+    # get the interface kind
+    linkinfo = msg.get_attr('IFLA_LINKINFO')
+    if linkinfo is not None:
+        kind = [x[1] for x in linkinfo['attrs']
+                if x[0] == 'IFLA_INFO_KIND']
+        if kind:
+            kind = kind[0]
+
+    if kind == 'tuntap':
+        return manage_tuntap(msg)
+    elif kind == 'team':
+        return manage_team(msg)
+    elif kind in ('ovs-bridge', 'openvswitch'):
+        return manage_ovs(msg)
+
+    if ANCIENT and kind in ('bridge', 'bond'):
+        # route the request
+        if kind == 'bridge':
+            compat_create_bridge(msg.get_attr('IFLA_IFNAME'))
+        elif kind == 'bond':
+            compat_create_bond(msg.get_attr('IFLA_IFNAME'))
+        # while RTM_NEWLINK is not intercepted -- sleep
+        time.sleep(_ANCIENT_BARRIER)
+        return
+
+    return {'verdict': 'forward',
+            'data': data}
+
+
+def manage_team(msg):
+
+    assert msg['header']['type'] == RTM_NEWLINK
+
+    config = {'device': msg.get_attr('IFLA_IFNAME'),
+              'runner': {'name': 'activebackup'},
+              'link_watch': {'name': 'ethtool'}}
+
+    with open(os.devnull, 'w') as fnull:
+        subprocess.check_call(['teamd', '-d', '-n', '-c',
+                               json.dumps(config)],
+                              stdout=fnull,
+                              stderr=fnull)
+
+
+def manage_team_port(cmd, master, ifname):
+    remap = {'add': 'add',
+             'del': 'remove'}
+    cmd = remap[cmd]
+    with open(os.devnull, 'w') as fnull:
+        subprocess.check_call(['teamdctl', master, 'port', cmd, ifname],
+                              stdout=fnull,
+                              stderr=fnull)
+    # do NOT forward request after
+    return False
+
+
+def manage_ovs(msg):
+    linkinfo = msg.get_attr('IFLA_LINKINFO')
+    ifname = msg.get_attr('IFLA_IFNAME')
+    kind = linkinfo.get_attr('IFLA_INFO_KIND')
+
+    # operations map
+    op_map = {RTM_NEWLINK: {'ovs-bridge': 'add-br',
+                            'openvswitch': 'add-br'},
+              RTM_DELLINK: {'ovs-bridge': 'del-br',
+                            'openvswitch': 'del-br'}}
+    op = op_map[msg['header']['type']][kind]
+
+    # make a call
+    with open(os.devnull, 'w') as fnull:
+        subprocess.check_call(['ovs-vsctl', op, ifname],
+                              stdout=fnull,
+                              stderr=fnull)
+
+
+def manage_tuntap(msg):
+
+    if TUNSETIFF is None:
+        raise Exception('unsupported arch')
+
+    if msg['header']['type'] != RTM_NEWLINK:
+        raise Exception('unsupported event')
+
+    ifru_flags = 0
+    linkinfo = msg.get_attr('IFLA_LINKINFO')
+    infodata = linkinfo.get_attr('IFLA_INFO_DATA')
+
+    flags = infodata.get_attr('IFTUN_IFR', None)
+    if infodata.get_attr('IFTUN_MODE') == 'tun':
+        ifru_flags |= IFT_TUN
+    elif infodata.get_attr('IFTUN_MODE') == 'tap':
+        ifru_flags |= IFT_TAP
+    else:
+        raise ValueError('invalid mode')
+    if flags is not None:
+        if flags['no_pi']:
+            ifru_flags |= IFT_NO_PI
+        if flags['one_queue']:
+            ifru_flags |= IFT_ONE_QUEUE
+        if flags['vnet_hdr']:
+            ifru_flags |= IFT_VNET_HDR
+        if flags['multi_queue']:
+            ifru_flags |= IFT_MULTI_QUEUE
+    ifr = msg.get_attr('IFLA_IFNAME')
+    if len(ifr) > IFNAMSIZ:
+        raise ValueError('ifname too long')
+    ifr += (IFNAMSIZ - len(ifr)) * '\0'
+    ifr = ifr.encode('ascii')
+    ifr += struct.pack('H', ifru_flags)
+
+    user = infodata.get_attr('IFTUN_UID')
+    group = infodata.get_attr('IFTUN_GID')
+    #
+    fd = os.open(TUNDEV, os.O_RDWR)
+    try:
+        ioctl(fd, TUNSETIFF, ifr)
+        if user is not None:
+            ioctl(fd, TUNSETOWNER, user)
+        if group is not None:
+            ioctl(fd, TUNSETGROUP, group)
+        ioctl(fd, TUNSETPERSIST, 1)
+    except Exception:
+        raise
+    finally:
+        os.close(fd)
+
+
+def compat_create_bridge(name):
+    with open(os.devnull, 'w') as fnull:
+        subprocess.check_call(['brctl', 'addbr', name],
+                              stdout=fnull,
+                              stderr=fnull)
+
+
+def compat_create_bond(name):
+    with open(_BONDING_MASTERS, 'w') as f:
+        f.write('+%s' % (name))
+
+
+def compat_set_bond(name, cmd, value):
+    # FIXME: join with bridge
+    # FIXME: use internal IO, not bash
+    t = 'echo %s >/sys/class/net/%s/bonding/%s'
+    with open(os.devnull, 'w') as fnull:
+        return subprocess.call(['bash', '-c', t % (value, name, cmd)],
+                               stdout=fnull,
+                               stderr=fnull)
+
+
+def compat_set_bridge(name, cmd, value):
+    t = 'echo %s >/sys/class/net/%s/bridge/%s'
+    with open(os.devnull, 'w') as fnull:
+        return subprocess.call(['bash', '-c', t % (value, name, cmd)],
+                               stdout=fnull,
+                               stderr=fnull)
+
+
+def compat_del_bridge(name):
+    with open(os.devnull, 'w') as fnull:
+        subprocess.check_call(['ip', 'link', 'set',
+                               'dev', name, 'down'])
+        subprocess.check_call(['brctl', 'delbr', name],
+                              stdout=fnull,
+                              stderr=fnull)
+
+
+def compat_del_bond(name):
+    subprocess.check_call(['ip', 'link', 'set',
+                           'dev', name, 'down'])
+    with open(_BONDING_MASTERS, 'w') as f:
+        f.write('-%s' % (name))
+
+
+def compat_bridge_port(cmd, master, port):
+    if not ANCIENT:
+        return True
+    with open(os.devnull, 'w') as fnull:
+        subprocess.check_call(['brctl', '%sif' % (cmd), master, port],
+                              stdout=fnull,
+                              stderr=fnull)
+
+
+def compat_bond_port(cmd, master, port):
+    if not ANCIENT:
+        return True
+    remap = {'add': '+',
+             'del': '-'}
+    cmd = remap[cmd]
+    with open(_BONDING_SLAVES % (master), 'w') as f:
+        f.write('%s%s' % (cmd, port))
+
+
+def compat_get_master(name):
+    f = None
+
+    for i in (_BRIDGE_MASTER, _BONDING_MASTER):
+        try:
+            f = open(i % (name))
+            break
+        except IOError:
+            pass
+
+    if f is not None:
+        master = int(f.read())
+        f.close()
+        return master
+
+
+def get_interface_type(name):
+    '''
+    Utility function to get interface type.
+
+    Unfortunately, we can not rely on RTNL or even ioctl().
+    RHEL doesn't support interface type in RTNL and doesn't
+    provide extended (private) interface flags via ioctl().
+
+    Args:
+        * name (str): interface name
+
+    Returns:
+        * False -- sysfs info unavailable
+        * None -- type not known
+        * str -- interface type:
+            * 'bond'
+            * 'bridge'
+    '''
+    # FIXME: support all interface types? Right now it is
+    # not needed
+    try:
+        ifattrs = os.listdir('/sys/class/net/%s/' % (name))
+    except OSError as e:
+        if e.errno == 2:
+            return 'unknown'
+        else:
+            raise
+
+    if 'bonding' in ifattrs:
+        return 'bond'
+    elif 'bridge' in ifattrs:
+        return 'bridge'
+    else:
+        return 'unknown'