+++ pyroute2/netlink/rtnl/req.py
@@ -1,7 +1,6 @@
 from socket import AF_INET6
 from pyroute2.common import basestring
-from pyroute2.netlink.rtnl.brmsg import brmsg
-from pyroute2.netlink.rtnl.bomsg import bomsg
+from pyroute2.netlink.rtnl.rtmsg import rtmsg
 
 
 class IPRequest(dict):
@@ -28,19 +27,26 @@ class IPRouteRequest(IPRequest):
         if isinstance(value, basestring) and value.find(':') >= 0:
             self['family'] = AF_INET6
         # work on the rest
-        if (key == 'dst') and (value != 'default'):
-            value = value.split('/')
-            if len(value) == 1:
-                dst = value[0]
-                mask = 0
-            elif len(value) == 2:
-                dst = value[0]
-                mask = int(value[1])
-            else:
-                raise ValueError('wrong destination')
-            dict.__setitem__(self, 'dst', dst)
-            dict.__setitem__(self, 'dst_len', mask)
-        elif key != 'dst':
+        if key == 'dst':
+            if value != 'default':
+                value = value.split('/')
+                if len(value) == 1:
+                    dst = value[0]
+                    mask = 0
+                elif len(value) == 2:
+                    dst = value[0]
+                    mask = int(value[1])
+                else:
+                    raise ValueError('wrong destination')
+                dict.__setitem__(self, 'dst', dst)
+                dict.__setitem__(self, 'dst_len', mask)
+        elif key == 'metrics':
+            ret = {'attrs': []}
+            for name in value:
+                rtax = rtmsg.metrics.name2nla(name)
+                ret['attrs'].append([rtax, value[name]])
+            dict.__setitem__(self, 'metrics', ret)
+        else:
             dict.__setitem__(self, key, value)
 
 
@@ -64,16 +70,6 @@ class CBRequest(IPRequest):
             dict.__setitem__(self, key, value)
 
 
-class BridgeRequest(CBRequest):
-    commands = [brmsg.nla2name(i[0]) for i in brmsg.commands.nla_map]
-    msg = brmsg
-
-
-class BondRequest(CBRequest):
-    commands = [bomsg.nla2name(i[0]) for i in bomsg.commands.nla_map]
-    msg = bomsg
-
-
 class IPLinkRequest(IPRequest):
     '''
     Utility class, that converts human-readable dictionary
@@ -83,6 +79,7 @@ class IPLinkRequest(IPRequest):
                  'carrier_changes']
 
     def __init__(self, *argv, **kwarg):
+        self.deferred = []
         IPRequest.__init__(self, *argv, **kwarg)
         if 'index' not in self:
             self['index'] = 0
@@ -105,20 +102,62 @@ class IPLinkRequest(IPRequest):
 
         # set up specific keys
         if key == 'kind':
-            if 'IFLA_LINKINFO' not in self:
-                self['IFLA_LINKINFO'] = {'attrs': []}
-            nla = ['IFLA_INFO_KIND', value]
-            # FIXME: we need to replace, not add
-            self['IFLA_LINKINFO']['attrs'].append(nla)
+            self['IFLA_LINKINFO'] = {'attrs': []}
+            linkinfo = self['IFLA_LINKINFO']['attrs']
+            linkinfo.append(['IFLA_INFO_KIND', value])
+            if value in ('vlan', 'bond', 'tuntap', 'veth'):
+                linkinfo.append(['IFLA_INFO_DATA', {'attrs': []}])
         elif key == 'vlan_id':
-            if 'IFLA_LINKINFO' not in self:
-                self['IFLA_LINKINFO'] = {'attrs': []}
-            nla = ['IFLA_INFO_DATA', {'attrs': [['IFLA_VLAN_ID', value]]}]
+            nla = ['IFLA_VLAN_ID', value]
             # FIXME: we need to replace, not add
-            self['IFLA_LINKINFO']['attrs'].append(nla)
-        elif key == 'bond_mode':
-            if 'IFLA_LINKINFO' not in self:
-                self['IFLA_LINKINFO'] = {'attrs': []}
-            nla = ['IFLA_INFO_DATA', {'attrs': [['IFLA_BOND_MODE', value]]}]
-            self['IFLA_LINKINFO']['attrs'].append(nla)
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'vlan')
+        elif key == 'gid':
+            nla = ['IFTUN_UID', value]
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'tuntap')
+        elif key == 'uid':
+            nla = ['IFTUN_UID', value]
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'tuntap')
+        elif key == 'mode':
+            nla = ['IFTUN_MODE', value]
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'tuntap')
+            nla = ['IFLA_BOND_MODE', value]
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'bond')
+        elif key == 'ifr':
+            nla = ['IFTUN_IFR', value]
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'tuntap')
+        elif key == 'peer':
+            nla = ['VETH_INFO_PEER', {'attrs': [['IFLA_IFNAME', value]]}]
+            self.defer_nla(nla, ('IFLA_LINKINFO', 'IFLA_INFO_DATA'),
+                           lambda x: x.get('kind', None) == 'veth')
         dict.__setitem__(self, key, value)
+        if self.deferred:
+            self.flush_deferred()
+
+    def flush_deferred(self):
+        deferred = []
+        for nla, path, predicate in self.deferred:
+            if predicate(self):
+                self.append_nla(nla, path)
+            else:
+                deferred.append((nla, path, predicate))
+        self.deferred = deferred
+
+    def append_nla(self, nla, path):
+            pwd = self
+            for step in path:
+                if step in pwd:
+                    pwd = pwd[step]
+                else:
+                    pwd = [x[1] for x in pwd['attrs']
+                           if x[0] == step][0]['attrs']
+            pwd.append(nla)
+
+    def defer_nla(self, nla, path, predicate):
+        self.deferred.append((nla, path, predicate))
+        self.flush_deferred()